// lib/mock-data/types.ts
export type UserRole = "CONSUMER" | "BUSINESS_OWNER" | "ADMIN";
// ...
export interface MockUser {
  id: string;
  email: string;
  password: string;
  name: string | null;
  phone: string | null;
  role: UserRole;
  image: string | null;
  emailVerified: Date | null;
  createdAt: Date;
  updatedAt: Date;
}